--- customer/src/tools/functions/functions.py
@@ -1,16 +1,0 @@
-
-def GetCustomerDetails(customer_id):
-    print('getting customer details')
-    print("Received keyword arguments:", customer_id)
-
-def GetCustomerMetadata(customer_id):
-    print('getting customer metadata')
-    print("Received keyword arguments:", customer_id)
-
-def GetCustomerMetrics(customer_id):
-    print('getting customer retentiion metrics')
-    print("Received keyword arguments:", customer_id)
-
-def GetCustomerPaymentInformation(customer_id):
-    print('getting customer Payments')
-    print("Received keyword arguments:", customer_id)
