+++ customer/src/tools/requests.py
@@ -1,0 +1,8 @@
+
+import requests
+
+def get(url:str):
+    # GET request to GeeksforGeeks
+    get_response = requests.get(url)
+
+    return get_response
